+++ builds/4chan-X.js
@@ -110,7 +110,8 @@
         'Hide Original Post Form': [true, 'Hide the normal post form.'],
         'Cooldown': [true, 'Indicate the remaining time before posting again.'],
         'Cooldown Prediction': [true, 'Decrease the cooldown time by taking into account upload speed. Disable it if it\'s inaccurate for you.'],
-        'Mark own Posts': [true, 'Mark posts you own.']
+        'Mark own Posts': [true, 'Mark posts you own.'],
+        'Tab to Choose Files First': [false, 'Tab to the file input before the submit button.']
       },
       'Quote Links': {
         'Quote Backlinks': [true, 'Add quote backlinks.'],
@@ -4969,6 +4970,9 @@
         status: $('[type=submit]', dialog),
         fileInput: $('[type=file]', dialog)
       };
+      if (Conf['Tab to Choose Files First']) {
+        $.add(nodes.fileSubmit, nodes.status);
+      }
       mimeTypes = $('ul.rules > li').textContent.trim().match(/: (.+)/)[1].toLowerCase().replace(/\w+/g, function(type) {
         switch (type) {
           case 'jpg':