+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Added the option `Tab to Choose Files First`, disabled by default.
+
 ## 3.4.0 - *2013-05-07*
 
 - **New feature**: `Archive selection`